GBE: define python interpreter by cmake variable
In some distros, python is linked to python3 not
python2, and GBE can't be built on such distros
without modification.

CMake provides a variable PYTHON_EXECUTABLE.
By default, this variable is the same as
`/usr/bin/env python`, and if another python2
interpreter is needed, just add this defination in
`cmake` command.

-DPYTHON_EXECUTABLE:FILEPATH=/path/to/python2

And this will change PYTHON_EXECUTABLE to
/path/to/python2

Signed-off-by: Boqun Feng <boqun.feng@gmail.com>
Reviewed-by: Zhigang Gong <zhigang.gong@linux.intel.com>