CL/Runtime: workaround the unused sampler_t kernel argument.
Current implementation is to use a normal integer to represent
a sampler_t, then later when the sampler is used in read_image
or get_sampler_info, the backend will fixup its type to SAMPLER.

But some test case in piglit will define a sampler_t kernel argument
with an empty kernel budy. Then we will not have a chance to fixup
the kernel argument type to sampler, then we will fail at runtime side.

To workaround this issue, we change the sampler_t to short type.
Then when the user call clSetKernelArg to set a sampler, it will pass
in a pointer size with a short value argument type. It will fail
the size checking logic, then we fixup its type to sampler there.

As this workaround will only take effect when error occur, it will
not bring too much side effect to the normal cases. And it can
pass the existing test cases.

Signed-off-by: Zhigang Gong <zhigang.gong@intel.com>
Reviewed-by: Yang Rong <rong.r.yang@intel.com>